HOTFIX : last byte in file not copied properly + remove useless line
MakePAK would not properly copy the last byte in a file. This was not
apparent initially because up until now I was only using it on textual
data files. Now that I began using it with graphics, the bug became
horrendesly apparent.

Furthermore, I deleted a completely useless line in the code. Yes, I
know they are supposed to be in separate commits, but I have my reasons,
and I will never tell them to you! I am just that evil! =)

Uh...this commit message is pretty long....